fixed issue where last word was lower case, fixed issue with numbers in the title
