+++ unit_test/test_validate_workflow.py
@@ -1,10 +1,6 @@
 import unittest
 from icon_validator.validate import validate
 
-# Import plugin validators to pass to tests
-from icon_validator.rules.plugin_validators.title_validator import TitleValidator
-from icon_validator.rules.plugin_validators.profanity_validator import ProfanityValidator
-
 # Import workflow validators to pass to tests
 from icon_validator.rules.workflow_validators.workflow_profanity_validator import WorkflowProfanityValidator
 from icon_validator.rules.workflow_validators.workflow_vendor_validator import WorkflowVendorValidator
@@ -19,39 +15,6 @@ from icon_validator.rules.workflow_validators.workflow_icon_filename_validator i
 from icon_validator.rules.workflow_validators.workflow_screenshot_validator import WorkflowScreenshotValidator
 from icon_validator.rules.workflow_validators.workflow_title_validator import WorkflowTitleValidator
 from icon_validator.rules.workflow_validators.workflow_description_validator import WorkflowDescriptionValidator
-
-
-class TestPluginValidate(unittest.TestCase):
-
-    def test_plugin_validate(self):
-        # example workflow in plugin_examples directory. Run tests with these files
-        directory_to_test = "plugin_examples/good_test"
-        file_to_test = "plugin.spec.yaml"
-        result = validate(directory_to_test, file_to_test, False, True)
-        self.assertFalse(result)
-
-    def test_title_validator(self):
-        # example workflow in plugin_examples directory. Run tests with these files
-        directory_to_test = "plugin_examples/title_tests"
-        file_to_test = "plugin_no_title.spec.yaml"
-        result = validate(directory_to_test, file_to_test, False, True, [TitleValidator()])
-        self.assertTrue(result)
-
-    def test_profanity_validator(self):
-        # example workflow in plugin_examples directory. Run tests with these files
-        directory_to_test = "plugin_examples/profanity_tests"
-        file_to_test = "plugin.spec.yaml"
-        result = validate(directory_to_test, file_to_test, False, True, [ProfanityValidator()])
-        self.assertTrue(result)
-
-    def test_plugin_with_false_for_required_on_output(self):
-        # TODO This validator is not correctly made: fix
-        # example workflow in plugin_examples directory. Run tests with these files
-        directory_to_test = "plugin_examples/bad_plugin_no_required_key_in_output"
-        file_to_test = "plugin.spec.yaml"
-        result = validate(directory_to_test, file_to_test, False, True)
-        self.assertTrue(result)
-
 
 class TestWorkflowValidate(unittest.TestCase):
